+++ lib/flutter_regex.dart
@@ -3,3 +3,5 @@ library flutter_regex;
 export 'package:flutter_regex/src/flutter_regex_pattern.dart';
 export 'package:flutter_regex/src/flutter_regex_validator.dart';
 export 'package:flutter_regex/src/flutter_regex_extension.dart';
+export 'package:flutter_regex/src/flutter_regex_email_extension.dart';
+export 'package:flutter_regex/src/flutter_regex_country.dart';